last changes

+++ src/com/itsonin/service/DeviceService.java
@@ -1,11 +1,19 @@
 package com.itsonin.service;
 
+import java.util.Date;
 import java.util.List;
+import java.util.UUID;
 
 import com.google.inject.Inject;
+import com.googlecode.objectify.Key;
+import com.itsonin.dao.CounterDao;
 import com.itsonin.dao.DeviceDao;
 import com.itsonin.entity.Device;
+import com.itsonin.enums.DeviceLevel;
+import com.itsonin.enums.SortOrder;
+import com.itsonin.exception.ForbiddenException;
 import com.itsonin.exception.NotFoundException;
+import com.itsonin.exception.UnauthorizedException;
 import com.itsonin.security.AuthContextService;
 
 /**
@@ -15,34 +23,67 @@ import com.itsonin.security.AuthContextService;
 public class DeviceService {
 
 	private DeviceDao deviceDao;
+	private CounterDao counterDao;
 	private AuthContextService authContextService;
 	
 	@Inject
-	public DeviceService(DeviceDao deviceDao, AuthContextService authContextService){
+	public DeviceService(DeviceDao deviceDao, CounterDao counterDao,
+			AuthContextService authContextService){
 		this.deviceDao = deviceDao;
+		this.counterDao = counterDao;
 		this.authContextService = authContextService;
 	}
 	
 	public Device create(Device device) {
-		deviceDao.save(device);
-		return device;
+		device.setDeviceId(counterDao.next("DEVICE"));
+		device.setCreated(new Date());
+		device.setToken(UUID.randomUUID().toString());//temporary TODO: replace
+		Key<Device> key = deviceDao.save(device);
+		return deviceDao.get(key);
 	}
 	
-	public List<Device> list() {
-		return deviceDao.list();
+	public List<Device> list(String name, Date created, Date lastLogin, String sortField,
+			 SortOrder sortOrder, Long offset, Long limit) {
+		if(!isAllowed())
+			throw new ForbiddenException("Not allowed");
+		
+		return deviceDao.list(name, created, lastLogin, sortField, sortOrder, offset, limit);
 	}
 	
 	public void delete(Long id) {
-		Device device = getDevice(id);
+		if(!isAllowed())
+			throw new ForbiddenException("Not allowed");
+		
+		Device device = get(id);
 		deviceDao.delete(device);
 	}
 	
-	private Device getDevice(Long id){
+	public Device getDeviceByToken(String token){
+		return deviceDao.getDeviceByToken(token);
+	}
+	
+	public Device authenticate(Long id){
+		Device device = deviceDao.get(id);
+		if(device == null)
+			throw new UnauthorizedException("Not autheticated");
+		else
+			return device;
+	}
+	
+	public Device get(Long id){
 		Device device = deviceDao.get(id);
 		if(device == null)
 			throw new NotFoundException("Device with id=" + id + " is not exists");
 		else
 			return device;
+	}
+	
+	boolean isAllowed(){
+		Device device = authContextService.get().getDevice();
+		if(DeviceLevel.SUPER.equals(device.getLevel()))
+			return true;
+		else
+			return false;
 	}
 	
 }
